problems

+++ KULENDAR/app/src/main/java/com/example/kulendar/DB/Alarm.kt
@@ -1,5 +1,5 @@
 package com.example.kulendar.DB
 
-data class Alarm(var notification_ID:Long ,var date:String, var title:String, var repeatOnOff:Int){
+data class Alarm(var notification_ID:Int ,var date:String, var title:String, var repeatOnOff:Int){
 
 }
